[Refactor] moved app logic to its own file and separated from server initialization
