+++ package.json
@@ -22,6 +22,8 @@
     "faker": "^4.1.0",
     "jest": "^23.6.0",
     "morgan": "^1.9.1",
+    "multer": "^1.4.1",
+    "multer-s3": "^2.9.0",
     "pg": "^7.6.1",
     "pg-promise": "^8.5.2",
     "supertest": "^3.3.0"
