+++ server/index.js
@@ -1,20 +1,7 @@
-const express = require('express');
-const app = express();
-const morgan = require('morgan');
-const bodyParser = require('body-parser');
-const db = require('../db');
-const startupDebugger = require('debug')('app:startup');
-const reviews = require('../router');
+const app = require('./app');
+
+//setup env variavles
 require('dotenv').config();
-
-//Middleware
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({extended: true}));
-app.use(morgan('tiny'));
-app.use(express.static(__dirname + '../client/dist'));
-
-//Routing
-app.use('/reviews', reviews);
 
 const port = process.env.PORT || 3003;
 const server = app.listen(port, () => {
